+++ workflows_definition/RELEASES.md
@@ -203,3 +203,11 @@ Based on:
 - Speakeasy CLI 1.446.0 (2.462.0) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [python v1.11.0] workflows_definition
+
+## 2024-11-28 00:31:17
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.api.epilot.io/workflows-definition.yaml
+- Speakeasy CLI 1.448.2 (2.467.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.11.1] workflows_definition